Add `lockTerrainZoom` animation option.
`_finalizeElevation` is usually called at the end of an animation and
it recalculates the zoom based on terrain data. For rotation around a
fixed centre this may change the zoom level because the elevation below
the rotating camera changes. What is desired though is for the zoom
level to stay constant. Hence the new option `lockTerrainZoom` which
inhibits the recalculation of a zoom level in `_finalizeElevation`. This
option is implicitly set in `rotateTo`, `resetNorth` and
`resetNorthPitch`.